fix: various thrashing against build/CI, squashed together
ci: additional build dependencies for CI gh-37

fix: fix some ifdef oops; hopefully this will work... gh-37

fix: another ifdef oops  gh-37

fix: oops, I already had an #if block for this all along

build: fixes for the cmake build gh-37

build: not finding statgrab.h on macOS -- maybe needs some extra CFLAGS
gh-37

fix: try to fix up some minor portability warnings gh-37

build: yet another cmake fix (worked for me locally but needed on homebrew) gh-37

docs: minor fixes for documentation build; still need to fix how
hawkmoth handles typedef struct ...  gh-37

build: more issues with libstatgrab on macOS with cmake

pray for me that this way works...

gh-37

build: should be just STATGRAB gh-37

build: ah, the issue now is actually in automake; maybe this will do it
gh-37

ci: add libstatgrab-dev and others for the doc builds gh-37

+++ include/asdf/core/ndarray.h
@@ -125,7 +125,7 @@ struct asdf_datatype {
     const char *name;
     asdf_byteorder_t byteorder;
     uint32_t ndim;
-    const size_t *shape;
+    const uint64_t *shape;
     uint32_t nfields;
     const asdf_datatype_t *fields;
 };
